Replace lockCanvas with lockHardwareCanvas in SimpleWebView
After replacing, the Surface encounter with an issue in Android Q:

	W OpenGLRenderer: dequeueBuffer failed, error = -110; switching to fallback

I guess that in Android Q (Android 10), the setFrameAvailableHandler should better be created in
the GLThread in which `Looper.myLooper()` should not be null.

However, in the GLThread attached to GLSurfaceView, `Looper.myLooper()` is null.

In fact, I build this library as a plugin for Unity. In Unity, I left setFrameAvailableHandler just be null.
It seems the issue does not happen again. So I'll just left it as it is.